Hamamatsu streak camera driver: a first working version
The streak camera is operated via the HPDTA software provided by Hamamatsu. The RemoteEx client provided by Hamamatsu allows to communicate to HPDTA.
The driver communicates with RemoteEx via the command- and the dataport.
It has four classes + dataflow class. The parent class is the streakcamera class, which has three children: readoutcam (to receive images), streakunit and delay generator.
    The most important methods are:
    - sendCommand: to send a command to RemoteEx
    - receiver thread: listening to response of RemoteEx
    - data thread: listening for images arriving or scaling tables to arrive