ASM wrapper: update MD every time a new field image is acquired
If we update the MD only in the start of a megafield, each dataarray (field) returned would have the same stage position. Update the MD every time a new field is acquired, so that the dataarray returned has the correct stage position.

TODO: A proper fix would be to have a settingsMD which includes MD such as the dwell time. And in the next call we would call mergeMD and add the settingsMD via updateMD call. Thus, we make sure that the generic settings such as the dwell time is for all fields the one that was set in the beginning of the megafield acquisition, but the stage position is updated correctly per field.