technolution driver: add hack to skip the first field image
The current ASM HW does not scan the very first field image correctly. This issue needs to be fixed in HW. However, until this is done, we need to "through away" the first field image and scan it a second time to receive a good first field image. To do so, just always scan the first image twice and overwrite on disc (external storage).